requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
gTTS==2.2.2
idna==2.10
image==1.5.33
Pillow==8.2.0
playsound==1.2.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
